+++ docs/userguide/plotting/plotting.py
@@ -161,7 +161,7 @@ _ = dataset.plot()
 # If you prefer not using colormap, `cmap=None` should be used. For instance:
 
 # %%
-_ = dataset.plot(cmap=None)
+_ = dataset.plot(cmap=None, colorbar=False)
 
 # %% [markdown]
 # Note that, by default, **sans-serif** font are used for all text in the figure.
@@ -374,15 +374,31 @@ prefs.axes_facecolor = "white"
 _ = dataset.plot_image(colorbar=True)  # will use image_cmap preference!
 
 # %%
-# here we use the generic `plot()` with the `method' argument and change the image_cmap:
+# Here we use the generic `plot()` with the `method' argument and we change the image_cmap:
 _ = dataset.plot(method="image", image_cmap="jet", colorbar=True)
+
 # %%
+# The colormap normalization can be changed using the `norm` parameter, as illustrated below,
+# for a centered colomap:
+import matplotlib as mpl
+
+norm = mpl.colors.CenteredNorm()
+_ = dataset.plot(method="image", image_cmap="jet", colorbar=True, norm=norm)
+
+# %%
+# or below for a log scale (more information about colormap normalization can be found
+# [here](https://matplotlib.org/stable/users/explain/colors/colormapnorms.html)).
+norm = mpl.colors.LogNorm(vmin=0.1, vmax=4.0)
+_ = dataset.plot(method="image", image_cmap="jet", colorbar=True, norm=norm)
+# %%
+# Below an example of a waterfall plot:
 prefs.reset()
 _ = dataset.plot_waterfall(figsize=(7, 4), y_reverse=True)
 
 # %%
+# And finally an example of a surface plot:
 prefs.reset()
-_ = dataset.plot_surface(linewidth=0.00, antialiased=True, y_reverse=True)
+_ = dataset.plot_surface(linewidth=0, y_reverse=True, autolayout=False)
 # %% [markdown]
 # ## Plotting 1D datasets